Make most things compile again.
