+++ node/network/collator-protocol/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2018"
 
 [dependencies]
 always-assert = "0.1.2"
+derive_more = "0.99.14"
 futures = "0.3.15"
 futures-timer = "3"
 thiserror = "1.0.23"
