+++ node/network/dispute-distribution/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 [dependencies]
 futures = "0.3.15"
 tracing = "0.1.26"
+derive_more = "0.99.14"
 parity-scale-codec = { version = "2.0.0", features = ["std"]  }
 polkadot-primitives = { path = "../../../primitives" }
 polkadot-erasure-coding = { path = "../../../erasure-coding" }
