+++ node/network/statement-distribution/Cargo.toml
@@ -20,6 +20,7 @@ arrayvec = "0.5.2"
 indexmap = "1.7.0"
 parity-scale-codec = { version = "2.0.0", default-features = false, features = ["derive"] }
 thiserror = "1.0.23"
+derive_more = "0.99.11"
 
 [dev-dependencies]
 polkadot-node-subsystem-test-helpers = { path = "../../subsystem-test-helpers" }
