+++ node/network/statement-distribution/src/error.rs
@@ -18,8 +18,10 @@
 //! Error handling related code and Error/Result definitions.
 
 use polkadot_node_network_protocol::PeerId;
+use polkadot_node_subsystem_util::runtime;
 use polkadot_primitives::v1::{CandidateHash, Hash};
 use polkadot_subsystem::SubsystemError;
+
 use thiserror::Error;
 
 use crate::LOG_TARGET;
@@ -32,7 +34,8 @@ pub type NonFatalResult<T> = std::result::Result<T, NonFatal>;
 pub type FatalResult<T> = std::result::Result<T, Fatal>;
 
 /// Errors for statement distribution.
-#[derive(Debug, Error, From)]
+#[derive(Debug, Error, derive_more::From)]
+#[error(transparent)]
 pub enum Error {
 	/// Fatal errors of dispute distribution.
 	Fatal(Fatal),
@@ -108,8 +111,10 @@ pub enum NonFatal {
 pub fn log_error(result: Result<()>, ctx: &'static str) -> std::result::Result<(), Fatal> {
 	match result {
 		Err(Error::Fatal(f)) => Err(f),
-		Err(Error::NonFatal(e)) =>
-			tracing::warn!(target: LOG_TARGET, error = ?error, ctx),
+		Err(Error::NonFatal(error)) => {
+			tracing::warn!(target: LOG_TARGET, error = ?error, ctx);
+			Ok(())
+		},
 		Ok(()) => Ok(()),
 	}
 }
