+++ node/subsystem-util/Cargo.toml
@@ -17,6 +17,7 @@ rand = "0.8.3"
 thiserror = "1.0.23"
 tracing = "0.1.26"
 lru = "0.6.5"
+derive_more = "0.99.11"
 
 polkadot-node-primitives = { path = "../primitives" }
 polkadot-node-subsystem = { package = "polkadot-node-subsystem", path = "../subsystem" }
